refactor favoritesVC

+++ MovieApp/MovieApp/FavoritesViewController/View/FavoritesViewController.swift
@@ -13,6 +13,7 @@ class FavoritesViewController: UIViewController {
     //MARK:- Properties -
     private let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     var detailMovies: [MovieCoreData] = []
+    var viewModel: FavoritesViewModelType?
     
     //MARK:- IBOutlets
     @IBOutlet weak var favoritesTableView: UITableView! {
@@ -30,6 +31,7 @@ class FavoritesViewController: UIViewController {
     //MARK: - LifeCycle -
     override func viewDidLoad() {
         super.viewDidLoad()
+        viewModel = FavoritesViewModel()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -39,12 +41,7 @@ class FavoritesViewController: UIViewController {
     
     //MARK: - Private func -
     private func getData() {
-        do {
-            detailMovies = try context.fetch(MovieCoreData.fetchRequest())
-            favoritesTableView.reloadData()
-        } catch {
-            print("Fetching Failed")
-        }
+        viewModel?.getData()
     }
 }
 
@@ -52,27 +49,27 @@ class FavoritesViewController: UIViewController {
 extension FavoritesViewController: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         
-        return detailMovies.count
+        return viewModel?.numberOfRows() ?? 1
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: Cells.mainCellIdentefier.rawValue, for: indexPath) as! MainTableViewCell
-        cell.configureFromCoreData(detailMovies[indexPath.row])
+        
+        let cellViewModel = viewModel?.cellFavoritesViewModel(forIndexPath: indexPath)
+        
+        cell.favoritesViewModel = cellViewModel
+        
         cell.backgroundColor = .white
         return cell
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
-            let task = detailMovies[indexPath.row]
+            guard let task = viewModel?.detailMovies[indexPath.row] else { return  }
             context.delete(task)
+            viewModel?.detailMovies.remove(at: indexPath.row)
             (UIApplication.shared.delegate as! AppDelegate).saveContext()
-            
-            do {
-                detailMovies = try context.fetch(MovieCoreData.fetchRequest())
-            } catch {
-                print("Fetching Failed")
-            }
+         
         }
         favoritesTableView.reloadData()
     }
